# docs/swagger.yaml
basePath: /
definitions:
  api.Activity:
    properties:
      until:
        type: integer
      weeks:
        items:
          type: integer
        type: array
    type: object
...
  api.ColorShare:
    properties:
      color:
        type: string
      games:
        type: integer
      share:
        type: number
    type: object
  api.DiedWith:
    properties:
      deaths:
        type: integer
      games:
        type: integer
      impostorId:
        type: string
      rate:
        type: number
    type: object
  api.DuoWinrate:
    properties:
      games:
        type: integer
      teammateId:
        type: string
      userId:
        type: string
      winrate:
        type: number
      wins:
        type: integer
    type: object
...
  api.FirstTarget:
    properties:
      crewmateGames:
        description: CrewmateGames is how many games the player was a crewmate in,
          the denominator of Rate.
        type: integer
      firstDeaths:
        description: FirstDeaths is how many games this player was the first to die
          in.
        type: integer
      rate:
        type: number
      userId:
        type: string
    type: object
...
  api.GuildLeaderboards:
    properties:
      bestCrewmateDuo:
        items:
          $ref: '#/definitions/api.DuoWinrate'
        type: array
      bestImpostorDuo:
        description: |-
          The duo boards rank pairs of players who shared a role in a game. Each pair appears once, lower user ID
          first. Best is highest winrate first; worst is lowest first.
        items:
          $ref: '#/definitions/api.DuoWinrate'
        type: array
      crewmateWinrate:
        items:
          $ref: '#/definitions/api.PlayerWinrate'
        type: array
      firstTarget:
        description: FirstTarget ranks players by how often they were the first to
          die.
        items:
          $ref: '#/definitions/api.FirstTarget'
        type: array
      impostorWinrate:
        items:
          $ref: '#/definitions/api.PlayerWinrate'
        type: array
      killedBy:
        description: |-
          KilledBy ranks crewmate and impostor pairs by how often the crewmate died with that impostor in the game.
          The game never reports who made a kill, so a death counts against every impostor of that game.
        items:
          $ref: '#/definitions/api.KilledBy'
        type: array
      minGames:
        description: |-
          MinGames is the guild's leaderboard minimum: the games a player or crewmate duo needs to be ranked by
          rate. The impostor duo boards use a fixed floor of two shared games instead.
        type: integer
      mostGames:
        description: MostGames ranks players by games recorded, with no minimum.
        items:
          $ref: '#/definitions/api.PlayerGames'
        type: array
      winrate:
        description: Winrate ranks players by winrate across both roles.
        items:
          $ref: '#/definitions/api.PlayerWinrate'
        type: array
      worstCrewmateDuo:
        items:
          $ref: '#/definitions/api.DuoWinrate'
        type: array
      worstImpostorDuo:
        items:
          $ref: '#/definitions/api.DuoWinrate'
        type: array
    type: object
...
  api.GuildStats:
    properties:
      generatedAt:
        description: GeneratedAt is the Unix time the rollup was built; responses
          may be served from a short cache.
        type: integer
      guildId:
        type: string
      leaderboards:
        allOf:
        - $ref: '#/definitions/api.GuildLeaderboards'
        description: Leaderboards is omitted for guilds whose premium is free or expired.
      players:
        additionalProperties:
          $ref: '#/definitions/api.StatsPlayer'
        description: |-
          Players maps every user ID named in the leaderboards to a name and picture, resolved through Discord with
          the bot's credentials or from the names the bot cached. IDs nothing knows are absent and the page shows
          the ID itself.
        type: object
      premium:
        allOf:
        - $ref: '#/definitions/premium.PremiumRecord'
        description: |-
          Premium is the guild's premium status the rollup was built under, so the page can explain a missing
          leaderboards section without a second request.
      summary:
        $ref: '#/definitions/api.GuildStatsSummary'
    type: object
  api.GuildStatsSummary:
    properties:
      crewmateWinrate:
        type: number
      crewmateWins:
        type: integer
      gamesPlayed:
        type: integer
      impostorWinrate:
        type: number
      impostorWins:
        type: integer
    type: object
  api.HttpError:
    properties:
      error:
        type: string
      statusCode:
        type: integer
    type: object
...
  api.KilledBy:
    properties:
      deaths:
        description: Deaths is how many of the shared games the crewmate died in.
        type: integer
      games:
        description: Games is how many games the two shared as crewmate and impostor,
          the denominator of Rate.
        type: integer
      impostorId:
        type: string
      rate:
        type: number
      userId:
        type: string
    type: object
  api.MapRecord:
    properties:
      games:
        type: integer
      map:
        type: string
      winrate:
        type: number
      wins:
        type: integer
    type: object
...
  api.PlayerGames:
    properties:
      games:
        type: integer
      userId:
        type: string
    type: object
  api.PlayerWinrate:
    properties:
      games:
        type: integer
      userId:
        type: string
      winrate:
        type: number
      wins:
        type: integer
    type: object
...
  api.StatsPlayer:
    properties:
      avatar:
        description: |-
          Avatar is a Discord CDN image URL: the user's avatar for this guild, else their global avatar, else the
          default Discord assigns them. Always set when the profile was resolved through Discord.
        type: string
      globalName:
        description: GlobalName is the display name the user chose for all of Discord,
          if any.
        type: string
      nickname:
        description: Nickname is the user's name in this guild, if they set one.
        type: string
      username:
        type: string
    type: object
...
  api.Teammate:
    properties:
      games:
        type: integer
      userId:
        type: string
      winrate:
        type: number
      wins:
        type: integer
    type: object
  api.UserFates:
    properties:
      killedAsCrewmate:
        $ref: '#/definitions/api.Fate'
      votedOutAsCrewmate:
        $ref: '#/definitions/api.Fate'
      votedOutAsImpostor:
        $ref: '#/definitions/api.Fate'
    type: object
  api.UserFirstTarget:
    properties:
      crewmateGames:
        type: integer
      firstDeaths:
        type: integer
      rate:
        type: number
    type: object
  api.UserMatch:
    properties:
      color:
        type: string
      endTime:
        type: integer
      map:
        description: Map is absent for matches recorded before maps were stored.
        type: string
      matchId:
        type: string
      name:
        type: string
      result:
        description: Result uses the keys of MatchSummary.Result.
        type: string
      role:
        type: string
      startTime:
        type: integer
      won:
        type: boolean
    type: object
...
  api.UserStats:
    properties:
      details:
        allOf:
        - $ref: '#/definitions/api.UserStatsDetails'
        description: Details is omitted for guilds whose premium is free or expired.
      generatedAt:
        description: GeneratedAt is the Unix time the document was built; responses
          may be served from a short cache.
        type: integer
      guildId:
        type: string
      players:
        additionalProperties:
          $ref: '#/definitions/api.StatsPlayer'
        description: |-
          Players maps the player and every user ID named in Details to a name and picture, resolved as for
          GET /guild/stats.
        type: object
      premium:
        allOf:
        - $ref: '#/definitions/premium.PremiumRecord'
        description: Premium is the guild's premium status the document was built
          under.
      recentMatches:
        description: |-
          RecentMatches are the player's latest finished matches, newest first, at most ten. Their match IDs work
          with GET /guild/match.
        items:
          $ref: '#/definitions/api.UserMatch'
        type: array
      summary:
        $ref: '#/definitions/api.UserStatsSummary'
      userId:
        type: string
    type: object
  api.UserStatsDetails:
    properties:
      activity:
        $ref: '#/definitions/api.Activity'
      bestCrewmateTeammates:
        description: |-
          The teammate boards rank players who shared a role with the player. Best is highest winrate first;
          worst is lowest first.
        items:
          $ref: '#/definitions/api.Teammate'
        type: array
      bestImpostorTeammates:
        items:
          $ref: '#/definitions/api.Teammate'
        type: array
      colors:
        items:
          $ref: '#/definitions/api.ColorShare'
        type: array
      fates:
        $ref: '#/definitions/api.UserFates'
      firstTarget:
        allOf:
        - $ref: '#/definitions/api.UserFirstTarget'
        description: FirstTarget is how often the player was the first to die, of
          their crewmate games.
      killedBy:
        description: |-
          KilledBy ranks impostors by how often the player died, as a crewmate, with them in the game. The game
          never reports who made a kill, so a death counts against every impostor of that game.
        items:
          $ref: '#/definitions/api.DiedWith'
        type: array
      maps:
        description: Maps is the player's record per map, most played first, over
          the games whose map was recorded.
        items:
          $ref: '#/definitions/api.MapRecord'
        type: array
      minGames:
        description: |-
          MinGames is the guild's leaderboard minimum: the games the player needs to be ranked by winrate, and a
          crewmate teammate or impostor needs to be listed. Impostor teammates need two shared games instead.
        type: integer
      names:
        items:
          $ref: '#/definitions/api.NameShare'
        type: array
      playedWith:
        description: PlayedWith ranks the linked players the player shared the most
          games with.
        items:
          $ref: '#/definitions/api.PlayedWith'
        type: array
      ranks:
        $ref: '#/definitions/api.UserRanks'
      streaks:
        $ref: '#/definitions/api.Streaks'
      survival:
        allOf:
        - $ref: '#/definitions/api.Survival'
        description: Survival is how often the player was still alive at the end of
          their crewmate games.
      worstCrewmateTeammates:
        items:
          $ref: '#/definitions/api.Teammate'
        type: array
      worstImpostorTeammates:
        items:
          $ref: '#/definitions/api.Teammate'
        type: array
    type: object
  api.UserStatsSummary:
    properties:
      crewmate:
        $ref: '#/definitions/api.RoleRecord'
      firstGame:
        description: FirstGame and LastGame are the start times of the player's first
          and latest games; absent with no games.
        type: integer
      games:
        type: integer
      impostor:
        $ref: '#/definitions/api.RoleRecord'
      lastGame:
        type: integer
      winrate:
        type: number
      wins:
        type: integer
    type: object
...
  premium.PremiumRecord:
    properties:
      days:
        type: integer
      tier:
        $ref: '#/definitions/premium.Tier'
    type: object
  premium.Tier:
    enum:
    - 0
    - 1
    - 2
    - 3
    - 4
    - 5
    format: int32
    type: integer
    x-enum-varnames:
    - FreeTier
    - BronzeTier
    - SilverTier
    - GoldTier
    - TrialTier
    - SelfHostTier
...
info:
  contact: {}
  description: AutoMuteUs API, served independently of Discord bot shards.
  title: AutoMuteUs
  version: dev
paths:
...
  /guild/stats:
    get:
      consumes:
      - application/json
      description: |-
        The guild statistics page in one document. Every guild gets the summary (games played and each
        side's wins); guilds with active premium also get the leaderboards the /stats guild slash command
        shows, five entries per board, honouring the guild's leaderboard minimum. User IDs are resolved to the
        names the bot last cached where available. Responses are built at most once a minute per guild.
      parameters:
      - description: Guild ID
        in: query
        name: guildID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.GuildStats'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HttpError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.HttpError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HttpError'
      security:
      - BasicAuth: []
      - DiscordBearer: []
      summary: Get Guild Stats
      tags:
      - guild
...
  /guild/user:
    get:
      consumes:
      - application/json
      description: |-
        One player's statistics in the guild. Every guild gets the player's games, wins, and winrates overall
        and per role, and their ten latest matches; guilds with active premium also get the player's ranks
        on the guild boards, streaks, survival, how often they were killed or voted out, first-target
        rate, favorite colors and names, who they played with and won or lost with, the impostors they
        died with, games per week, and per-map records. Any member may read any player, as with the /stats
        user slash command. Responses may be up to a minute old.
      parameters:
      - description: Guild ID
        in: query
        name: guildID
        required: true
        type: string
      - description: User ID
        in: query
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.UserStats'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.HttpError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.HttpError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.HttpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.HttpError'
      security:
      - BasicAuth: []
      - DiscordBearer: []
      summary: Get Player Stats
      tags:
      - guild
...
securityDefinitions:
  BasicAuth:
    type: basic
  DiscordBearer:
    description: 'Discord OAuth access token: Bearer <token>; requires identify and
      guilds.'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"